// solar-proposal-app/src/generated/prisma/models/Lead.ts
export type LeadModel = runtime.Types.Result.DefaultSelection<Prisma.$LeadPayload>
// ...
export type $LeadPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "Lead"
  objects: {
    scanJob: Prisma.$ScanJobPayload<ExtArgs> | null
    user: Prisma.$UserPayload<ExtArgs>
    solarAnalysis: Prisma.$SolarAnalysisPayload<ExtArgs> | null
    proposals: Prisma.$ProposalPayload<ExtArgs>[]
    campaignLeads: Prisma.$CampaignLeadPayload<ExtArgs>[]
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    businessName: string
    address: string
    city: string | null
    state: string | null
    country: string | null
    placeId: string | null
    phone: string | null
    website: string | null
    businessType: string | null
    rating: number | null
    totalRatings: number | null
    lat: number | null
    lng: number | null
    status: $Enums.LeadStatus
    notes: string | null
    ownerName: string | null
    ownerEmail: string | null
    ownerPhone: string | null
    ownerAddress: string | null
    ownerLookupAt: Date | null
    roofScore: number | null
    urgencyScore: number | null
    solarScore: number | null
    roofScoreLabel: string | null
    animationUrl: string | null
    scanJobId: string | null
    userId: string
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["lead"]>
  composites: {}
}